service/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.35.95
botocore==1.35.99
celery==5.5.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
distro==1.9.0
Django==5.2
django-allauth==65.3.1
django-cors-headers==4.7.0
django-elasticsearch-dsl==8.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
kombu==5.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.5
openai==1.77.0
openai-clip==1.0.1
packaging==24.2
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
rest-framework-simplejwt==0.0.2
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
